// src/types/InstanceSyncStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type InstanceSyncStatus = {
  hasManifest: boolean;
  canResume: boolean;
  isComplete: boolean;
  bytesVerified: number;
  bytesTotal: number;
  filesVerified: number;
  filesTotal: number;
  percentComplete: number;
};